refactor: Update navigation and state management in `Deck`
*   Replaces the deprecated `DeckNavigator` and `containerState` with the new `DeckNavigator` and `SlideConfig` throughout the presentation module.
*   Updates `DeckWebImpl` (JS and Wasm) to use `skiPresentationSlides` and synchronizes slide navigation using `localStorage` and `DeckSyncState`.
*   Refactors the `Deck` entry point and `PresentationDeck` to accept `slides` and `navigator` as explicit parameters.
*   Simplifies `main.kt` for JVM by sharing a single `DeckNavigator` instance between the "Slides" and "Presenter Notes" windows.
*   Adds medium padding to `DeckScaffolding` in `PresentationDeck.kt`.
*   Cleans up unused imports and deprecated references across several files.
