+++ composeApp/src/commonMain/kotlin/ke/don/ski/presentation/PresentationDeck.kt
@@ -1,5 +1,7 @@
 package ke.don.ski.presentation
 
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
@@ -7,8 +9,10 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
 import io.github.donald_okara.components.frames.SkiFrame
 import ke.don.design.theme.AppTheme
+import ke.don.design.theme.dimens
 import ke.don.ski.domain.SlideConfig
 import ke.don.ski.navigation.DeckNavigator
 import ke.don.ski.presentation.deprecated.DeckMode
@@ -17,8 +21,8 @@ import ke.don.ski.presentation.ui.skiPresentationSlides
 @Composable
 fun PresentationDeck(
     deckMode: DeckMode,
-    slides: List<SlideConfig> = skiPresentationSlides(),
-    navigator: DeckNavigator = DeckNavigator(skiPresentationSlides()),
+    slides: List<SlideConfig>,
+    navigator: DeckNavigator,
     guidesFrame: SkiFrame
 ) {
     var darkMode by rememberSaveable(deckMode) { mutableStateOf(deckMode == DeckMode.Local) }
@@ -28,6 +32,8 @@ fun PresentationDeck(
     ) {
         Surface {
             DeckScaffolding(
+                modifier = Modifier
+                    .padding(MaterialTheme.dimens.mediumPadding),
                 switchTheme = { darkMode = !darkMode },
                 darkTheme = darkMode,
                 frame = guidesFrame,
